falcon-wrapper-service/requirements.txt
opencv-python==4.9.0.80
kafka-python==2.0.2
requests==2.31.0